requirements.txt
pandas 
numpy 
matplotlib 
seaborn 
h2o 
streamlit 
google-generativeai 
python-dotenv

